RUST-2294 Fold serde-tests into regular tests
