+++ .evergreen/check-rustfmt.sh
@@ -4,5 +4,3 @@ set -o errexit
 
 . ~/.cargo/env
 cargo +nightly fmt -- --check
-
-cd serde-tests && cargo +nightly fmt -- --check
