--- serde-tests/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "serde-tests"
-version = "0.1.0"
-authors = ["Kevin Yeh <kevinyeah@utexas.edu>"]
-edition = "2018"
-
-[features]
-default = []
-
-[dependencies]
-bson = { path = "..", features = ["uuid-1", "chrono-0_4", "serde", "serde_with-3", "serde_json-1"] }
-serde = { version = "1.0", features = ["derive"] }
-pretty_assertions = "0.6.1"
-hex = "0.4.2"
-serde_with = "3"
-chrono = "0.4"
-uuid = "1"
-
-[dev-dependencies]
-serde_json = { version = "1", features = ["preserve_order"] }
-rmp-serde = "0.15"
-base64 = "0.13.0"
-
-[lib]
-name = "serde_tests"
-path = "lib.rs"
-
-[[test]]
-name = "serde"
-path = "test.rs"
