--- serde-tests/rustfmt.toml
@@ -1,9 +1,0 @@
-combine_control_expr = false
-comment_width = 100
-condense_wildcard_suffixes = true
-format_strings = true
-normalize_comments = true
-use_try_shorthand = true
-wrap_comments = true
-imports_layout = "HorizontalVertical"
-imports_granularity = "Crate"
